eliminated panic when serving requests from empty registry
when generating an etag from a query that returns no results,
indexing into the byte slice caused that thread to panic.

now checking for a nil byte slice or a small byte slice in
handlers.go:46:getEtag()

on nil byte slice, return empty string for etag. on small
byte slice, don't truncate it, but return the whole thing.